修复内容 	1. 会员等级背景图去掉校验 	2. 当查询不到会员等级是，按无会员等级展示 	3. 删除商城首页冗余配置在index中的引用 	4. 换绑推广人时，计算上一个推广人的推广人数 	5. pc后台清除推广人时，更新推广人数量 	6. 保证砍价金额最小为0.01 	7. 修复商品删除时购物车关联删除 	8. 删除商品问题修复 	9. 运费模板——指定包邮，包邮数量类型修改 	10. 签到错误修复 	11. 修复我的优惠券只查询20条的问题 	12. 文章列表修复 	13. 拼团商品详情页数据统计显示问题修复 	14. PC后台，账户详情，持有优惠券列表修复 	15. 支付查询参数修复 	16. 修复过期优惠券可以重复领取 	17. 订单邮费切换地址重复计算修复 	18. 判断是否在指定包邮区域内 必须满足件数 + 金额 才能包邮 	19. 支付页面，切换tab，金额计算问题修复 	20. 物流模板新增、编辑——修复 	21. 去除线下邮费的影响 	22. 订单运费计算重写 	23. 下单页面到店自提合计金额不应该计算商品邮费 	24. 新人券领取后，部分使用时间为空——修复

+++ crmeb/src/main/java/com/zbkj/crmeb/express/model/ShippingTemplatesFree.java
@@ -48,7 +48,7 @@ public class ShippingTemplatesFree implements Serializable {
     private String title;
 
     @ApiModelProperty(value = "包邮件数")
-    private Integer number;
+    private BigDecimal number;
 
     @ApiModelProperty(value = "包邮金额")
     private BigDecimal price;
